requirements.txt
streamlit
gspread
google-auth
num2words